+++ core/commands/daemon_restart.go
@@ -4,9 +4,9 @@ import (
 	"os"
 	"time"
 
-	"github.com/pkg/errors"
 	"github.com/opensvc/om3/daemon/daemoncli"
 	"github.com/opensvc/om3/util/command"
+	"github.com/pkg/errors"
 )
 
 type (
@@ -28,8 +28,8 @@ func (t *CmdDaemonRestart) Run() error {
 		}
 	} else {
 		args := []string{"daemon", "restart"}
-		if t.Debug {
-			args = append(args, "--debug")
+		if t.Log != "" {
+			args = append(args, "--log", t.Log)
 		}
 		if t.Server != "" {
 			args = append(args, "--server", t.Server)