+++ core/commands/object_provision.go
@@ -37,9 +37,13 @@ func (t *CmdObjectProvision) Run(selector, kind string) error {
 		objectaction.WithRemoteAction("provision"),
 		objectaction.WithAsyncTarget("provisioned"),
 		objectaction.WithAsyncWatch(t.Watch),
-		objectaction.WithDigest(),
+		objectaction.WithProgress(!t.Quiet && t.Log == ""),
 		objectaction.WithLocalRun(func(ctx context.Context, p path.T) (interface{}, error) {
-			o, err := object.NewActor(p)
+			o, err := object.NewActor(p,
+				object.WithConsoleLog(t.Log != ""),
+				object.WithConsoleColor(t.Color != "no"),
+			)
+
 			if err != nil {
 				return nil, err
 			}