+++ core/commands/object_stop.go
@@ -35,9 +35,12 @@ func (t *CmdObjectStop) Run(selector, kind string) error {
 		objectaction.WithRemoteAction("stop"),
 		objectaction.WithAsyncTarget("stopped"),
 		objectaction.WithAsyncWatch(t.Watch),
-		objectaction.WithDigest(),
+		objectaction.WithProgress(!t.Quiet && t.Log == ""),
 		objectaction.WithLocalRun(func(ctx context.Context, p path.T) (interface{}, error) {
-			o, err := object.NewActor(p)
+			o, err := object.NewActor(p,
+				object.WithConsoleLog(t.Log != ""),
+				object.WithConsoleColor(t.Color != "no"),
+			)
 			if err != nil {
 				return nil, err
 			}