+++ core/commands/object_sync_update.go
@@ -32,8 +32,12 @@ func (t *CmdObjectSyncUpdate) Run(selector, kind string) error {
 		objectaction.WithColor(t.Color),
 		objectaction.WithRemoteNodes(t.NodeSelector),
 		objectaction.WithRemoteAction("sync update"),
+		objectaction.WithProgress(!t.Quiet && t.Log == ""),
 		objectaction.WithLocalRun(func(ctx context.Context, p path.T) (interface{}, error) {
-			o, err := object.NewActor(p)
+			o, err := object.NewActor(p,
+				object.WithConsoleLog(t.Log != ""),
+				object.WithConsoleColor(t.Color != "no"),
+			)
 			if err != nil {
 				return nil, err
 			}