+++ core/object/core.go
@@ -9,11 +9,13 @@ import (
 
 	"github.com/opensvc/om3/core/instance"
 	"github.com/opensvc/om3/core/path"
+	"github.com/opensvc/om3/core/rawconfig"
 	"github.com/opensvc/om3/core/xconfig"
 	"github.com/opensvc/om3/util/compliance"
 	"github.com/opensvc/om3/util/funcopt"
 	"github.com/opensvc/om3/util/hostname"
 	"github.com/opensvc/om3/util/logging"
+	"github.com/opensvc/om3/util/progress"
 	"github.com/opensvc/om3/util/xsession"
 )
 
@@ -25,8 +27,10 @@ type (
 		path path.T
 
 		// private
-		volatile bool
-		log      zerolog.Logger
+		volatile         bool
+		withConsoleLog   bool
+		withConsoleColor bool
+		log              zerolog.Logger
 
 		// caches
 		id         uuid.UUID
@@ -92,7 +96,8 @@ func (t *core) init(referrer xconfig.Referrer, id any, opts ...funcopt.O) error 
 		return err
 	}
 	t.log = logging.Configure(logging.Config{
-		ConsoleLoggingEnabled: zerolog.GlobalLevel() == zerolog.DebugLevel,
+		ConsoleLoggingEnabled: t.withConsoleLog,
+		ConsoleLoggingColor:   t.withConsoleColor,
 		EncodeLogsAsJSON:      true,
 		FileLoggingEnabled:    !t.volatile,
 		Directory:             t.logDir(), // contains the ns/kind
@@ -129,10 +134,8 @@ func (t *core) Path() path.T {
 	return t.path
 }
 
-//
 // ConfigFile returns the absolute path of an opensvc object configuration
 // file.
-//
 func (t core) ConfigFile() string {
 	if t.configFile == "" {
 		t.configFile = t.path.ConfigFile()
@@ -140,10 +143,8 @@ func (t core) ConfigFile() string {
 	return t.configFile
 }
 
-//
 // Node returns a cache Node struct pointer. If none is already cached,
 // allocate a new Node{} and cache it.
-//
 func (t *core) Node() (*Node, error) {
 	if t.node != nil {
 		return t.node, nil
@@ -158,4 +159,16 @@ func (t *core) Node() (*Node, error) {
 
 func (t core) Log() *zerolog.Logger {
 	return &t.log
+}
+
+func (t core) ProgressKey() []string {
+	p := rawconfig.Colorize.Bold(t.path.String())
+	return []string{p}
+}
+
+func (t core) Progress(ctx context.Context, cols ...any) {
+	if view := progress.ViewFromContext(ctx); view != nil {
+		key := t.ProgressKey()
+		view.Info(key, cols)
+	}
 }